ref(explore): Move get_pinned_attributes to event_search module
Move get_pinned_attributes and its helpers from the trace item stats
endpoint to sentry.api.event_search where other query parsing utilities
live. This makes the function discoverable and reusable, and keeps the
endpoint focused on API handling.
